Add automatic CCL list generation for prefill and decode when user does not provide lists (#663)
For using CCL feature during prefilling and decoding, the user needs to
pass two lists of context lengths to be used during these processes.
Here, we add an option to generate these lists automatically when they
are not provided by the user manually. These list generation is suitable
for a general-purpose application which considers both prefilling and
decoding processes and generates CCL lists for both.

---------

+++ QEfficient/transformers/models/gemma3/modeling_gemma3.py
@@ -921,7 +921,7 @@ class QEffGemma3ForConditionalGeneration(Gemma3ForConditionalGeneration):
         )
 
         if comp_ctx_lengths is not None:
-            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.long)
+            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int8)
         if continuous_batching:
             lang_inputs["batch_index"] = torch.arange(bs).view(bs, 1)
